// reel-tv/src/generated/prisma/enums.ts
/* !!! This is code generated by Prisma. Do not edit directly. !!! */
/* eslint-disable */
// biome-ignore-all lint: generated file
// @ts-nocheck 
/*
* This file exports all enum related types from the schema.
*
* 🟢 You can import this file directly.
*/

export const Role = {
  VIEWER: 'VIEWER',
  CREATOR: 'CREATOR',
  ADMIN: 'ADMIN'
} as const

export type Role = (typeof Role)[keyof typeof Role]


export const VideoStatus = {
  UPLOADING: 'UPLOADING',
  PROCESSING: 'PROCESSING',
  PENDING: 'PENDING',
  APPROVED: 'APPROVED',
  REJECTED: 'REJECTED',
  ACTIVE: 'ACTIVE',
  ARCHIVED: 'ARCHIVED'
} as const

export type VideoStatus = (typeof VideoStatus)[keyof typeof VideoStatus]


export const BroadcastPackage = {
  STARTER: 'STARTER',
  STANDARD: 'STANDARD',
  PRO: 'PRO',
  BUSINESS: 'BUSINESS'
} as const

export type BroadcastPackage = (typeof BroadcastPackage)[keyof typeof BroadcastPackage]


export const OrderStatus = {
  PENDING: 'PENDING',
  PAID: 'PAID',
  ACTIVE: 'ACTIVE',
  COMPLETED: 'COMPLETED',
  CANCELLED: 'CANCELLED',
  REFUNDED: 'REFUNDED'
} as const

export type OrderStatus = (typeof OrderStatus)[keyof typeof OrderStatus]


export const PaymentStatus = {
  PENDING: 'PENDING',
  SUCCEEDED: 'SUCCEEDED',
  FAILED: 'FAILED',
  REFUNDED: 'REFUNDED'
} as const

export type PaymentStatus = (typeof PaymentStatus)[keyof typeof PaymentStatus]


export const SubscriptionStatus = {
  ACTIVE: 'ACTIVE',
  CANCELLED: 'CANCELLED',
  PAST_DUE: 'PAST_DUE'
} as const

export type SubscriptionStatus = (typeof SubscriptionStatus)[keyof typeof SubscriptionStatus]


export const ReportStatus = {
  PENDING: 'PENDING',
  REVIEWING: 'REVIEWING',
  RESOLVED: 'RESOLVED',
  DISMISSED: 'DISMISSED'
} as const

export type ReportStatus = (typeof ReportStatus)[keyof typeof ReportStatus]


export const NotificationType = {
  ORDER_CONFIRMED: 'ORDER_CONFIRMED',
  VIDEO_APPROVED: 'VIDEO_APPROVED',
  VIDEO_REJECTED: 'VIDEO_REJECTED',
  NEW_SUBSCRIBER: 'NEW_SUBSCRIBER',
  PAYMENT_RECEIVED: 'PAYMENT_RECEIVED',
  SYSTEM: 'SYSTEM'
} as const

export type NotificationType = (typeof NotificationType)[keyof typeof NotificationType]


export const TransactionType = {
  PURCHASE: 'PURCHASE',
  REFUND: 'REFUND',
  PAYOUT: 'PAYOUT'
} as const

export type TransactionType = (typeof TransactionType)[keyof typeof TransactionType]


export const InvoiceStatus = {
  DRAFT: 'DRAFT',
  SENT: 'SENT',
  PAID: 'PAID',
  VOID: 'VOID',
  OVERDUE: 'OVERDUE'
} as const

export type InvoiceStatus = (typeof InvoiceStatus)[keyof typeof InvoiceStatus]